Bootstrapped the monthly dezuranje template.
Attempted to lower the number of required sql queries. Right now it's 1
query per dezuranje slot per day =~ 22 days * 3 slot = 66 queries, and
that's used just to see if someone had claimed that slot.  The claims need
to loaded once, and then partitioned up inside the controller, not with
separate db queries. Maybe that can be done with ORM caching in some way.